requirements.txt
aiohttp
asyncio
asyncpg
currency-exchange
discord.py
googletrans
jellyfish
pandas
pillow
psutil
py_expression_eval
python-dateutil
requests
stopwatch.py
testing.postgresql
